+++ include/le3d/core/assert.hpp
@@ -1,6 +1,7 @@
 #pragma once
+#include "le3d/defines.hpp"
 
-#if defined(DEBUGGING)
+#if defined(ASSERTS)
 #define ASSERT(predicate, errorMessage) le::assertMsg(!!(predicate), #errorMessage, __FILE__, __LINE__)
 #define ASSERT_VAR(predicate, szStr) le::assertMsg(!!(predicate), szStr, __FILE__, __LINE__)
 #else
