+++ include/le3d/game/utils.hpp
@@ -11,6 +11,31 @@ void renderMeshes(const HMesh& mesh, const std::vector<ModelMats>& mats, const H
 
 namespace debug
 {
+struct Quad2D
+{
+	std::optional<glm::vec4> oTexCoords;
+	std::optional<HMesh> oMesh;
+	glm::vec2 pos = glm::vec2(0.0f);
+	glm::vec2 size = glm::vec2(1.0f);
+	Colour tint = Colour::White;
+};
+
+struct Text2D
+{
+	enum class Align
+	{
+		Centre = 0,
+		Left,
+		Right
+	};
+
+	std::string text;
+	glm::vec2 pos = glm::vec2(0.0f);
+	f32 height = 40.0f;
+	Align align = Align::Centre;
+	Colour colour = Colour::White;
+};
+
 HMesh& debugCube();
 HMesh& debugQuad();
 HMesh& debugPyramid();
@@ -22,5 +47,11 @@ HMesh& debugCylinder();
 Model& debugArrow(const glm::quat& orientation);
 
 void unloadAll();
+
+void draw2DQuads(std::vector<Quad2D> quads, const HTexture& texture, const f32 uiAR = 0.0f);
+void renderString(const Text2D& text, const HFont& hFont, const f32 uiAR = 0.0f);
+
+extern Text2D g_fpsStyle;
+void renderFPS(const HFont& font, const f32 uiAR = 0.0f);
 } // namespace debug
 } // namespace le
